+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.0
 - Added `HueDecorator` using native `ColorFilter.matrix` for performance and safety.
+- added example
 
 # 1.0.0
 - Added collection of visual effect decorators.